requirements.txt
streamlit
datasets
python-dotenv